+++ src/munged/enc.c
@@ -267,8 +267,8 @@ enc_check_retry (munge_cred_t c)
 
     if (m->retry > 0) {
         log_msg (LOG_INFO,
-            "Encode retry #%d for client UID=%d GID=%d",
-            m->retry, m->client_uid, m->client_gid);
+            "Encode retry #%d for client UID=%u GID=%u", m->retry,
+            (unsigned int) m->client_uid, (unsigned int) m->client_gid);
     }
     if (m->retry > MUNGE_SOCKET_RETRY_ATTEMPTS) {
         return (m_msg_set_err (m, EMUNGE_SOCKET,
